+++ BeerProductionAPI/BeerProductionAPI/PersistenceLayer/MachineModule/IMachineReadData.cs
@@ -1,4 +1,5 @@
 ﻿using Opc.UaFx.Client;
+using System;
 
 namespace BeerProductionAPI
 {
@@ -14,7 +15,7 @@ namespace BeerProductionAPI
 
         int ReadCurrentState(OpcClient accessPoint);
 
-        int ReadBatchID(OpcClient accessPoint);
+        UInt16 ReadBatchID(OpcClient accessPoint);
 
         ushort ReadBatchSize(OpcClient accessPoint);
